User functionality, local time and favourite timezones added for user

+++ Global Time Zone/TimeZone.h
@@ -5,14 +5,14 @@
 class TimeZone {
 public:
     std::string name;
-    int utc_offset;
+    double utc_offset;
     bool observes_dst;
 
     // Default constructor
     TimeZone() : name("Unknown"), utc_offset(0), observes_dst(false) {}
 
     // Parameterized constructor
-    TimeZone(std::string n, int offset, bool dst)
+    TimeZone(std::string n, double offset, bool dst)
         : name(n), utc_offset(offset), observes_dst(dst) {
     }
 };
